Initial checkin
Basic button management working.  This uses a 3rd party Button
util as found on github now just to save some time.  The
"2nd" mouse button (what would be the right mouse button) is
what is acting as the mouse on/off button so I don't lose control
of my mouse on the computer in the event of a bug.